@@ -6,7 +6,7 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -15,14 +15,15 @@ dependencies:
     sdk: flutter         
   
   cupertino_icons: ^1.0.2
-  flutter_riverpod: ^0.14.0+3
-  dio: ^4.0.0
-  url_launcher: ^6.0.9
+  flutter_riverpod: ^2.3.6
+  dio: ^5.3.0
+  url_launcher: ^6.1.12
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_native_splash: ^1.2.1
+  flutter_lints: ^2.0.1
 
 
 
