@@ -9,26 +9,26 @@ class TwoColumnLayout extends StatefulWidget {
   final BoxConstraints constriants;
 
   @override
-  _TwoColumnLayoutState createState() => _TwoColumnLayoutState();
+  State<TwoColumnLayout> createState() => _TwoColumnLayoutState();
 }
 
 class _TwoColumnLayoutState extends State<TwoColumnLayout> {
   static const _verticalPaddindSmall = 10.0;
   static const _maxWhiteContainerWidth = 1280.0;
   static const _maxWhiteContainerHeight = 720.0;
-  double _verticalPadding = _verticalPaddindSmall;
+  final double _verticalPadding = _verticalPaddindSmall;
 
   @override
   Widget build(BuildContext context) {
-    final _maxWidth = widget.constriants.maxWidth;
-    final _maxHeight = widget.constriants.maxHeight;
+    final maxWidth = widget.constriants.maxWidth;
+    final maxHeight = widget.constriants.maxHeight;
     return Container(
-      width: _maxWidth,
-      height: _maxHeight,
+      width: maxWidth,
+      height: maxHeight,
       color: Theme.of(context).primaryColor,
       child: Center(
         child: AnimatedContainer(
-          constraints: BoxConstraints(
+          constraints: const BoxConstraints(
               maxHeight: _maxWhiteContainerHeight,
               maxWidth: _maxWhiteContainerWidth),
           decoration: BoxDecoration(
@@ -36,8 +36,8 @@ class _TwoColumnLayoutState extends State<TwoColumnLayout> {
             borderRadius: const BorderRadius.all(Radius.circular(16)),
           ),
           duration: const Duration(milliseconds: 10),
-          width: _maxWidth * 0.9,
-          height: _maxHeight * 0.95,
+          width: maxWidth * 0.9,
+          height: maxHeight * 0.95,
           child: Center(
             child: SingleChildScrollView(
               child: Padding(
@@ -49,24 +49,24 @@ class _TwoColumnLayoutState extends State<TwoColumnLayout> {
                     crossAxisAlignment: CrossAxisAlignment.stretch,
                     mainAxisSize: MainAxisSize.max,
                     children: [
-                      Row(
+                      const Row(
                         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                         children: [
-                          Expanded(child: const NameWidget()),
-                          const SizedBox(width: 8),
-                          Expanded(child: const AgeWidget())
+                          Expanded(child: NameWidget()),
+                          SizedBox(width: 8),
+                          Expanded(child: AgeWidget())
                         ],
                       ),
                       SizedBox(height: _verticalPadding),
-                      Consumer(builder: (context, watch, _) {
+                      Consumer(builder: (context, ref, _) {
                         return Row(
                           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                           children: [
-                            Expanded(child: const TypeOfIdDropdown()),
-                            if (watch(typeOfIdProvider).state != null)
-                              Expanded(
-                                  child:
-                                      const IdNumberWidget(isTwoColumn: true))
+                            const Expanded(child: TypeOfIdDropdown()),
+                            if (ref.watch(typeOfIdProvider) != null)
+                              const Expanded(
+                                child: IdNumberWidget(isTwoColumn: true),
+                              ),
                           ],
                         );
                       }),
@@ -77,16 +77,13 @@ class _TwoColumnLayoutState extends State<TwoColumnLayout> {
                         children: [
                           Expanded(
                             flex: 2,
-                            child: Container(
-                              child: Column(
-                                mainAxisAlignment:
-                                    MainAxisAlignment.spaceEvenly,
-                                children: [
-                                  const TypeOfTherapyDropDown(),
-                                  SizedBox(height: _verticalPadding),
-                                  const InsuranceWidget(),
-                                ],
-                              ),
+                            child: Column(
+                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                              children: [
+                                const TypeOfTherapyDropDown(),
+                                SizedBox(height: _verticalPadding),
+                                const InsuranceWidget(),
+                              ],
                             ),
                           ),
                           if (widget.constriants.maxWidth > 800)
@@ -101,43 +98,42 @@ class _TwoColumnLayoutState extends State<TwoColumnLayout> {
                         ],
                       ),
                       const SizedBox(height: 16),
-                      Row(
+                      const Row(
                         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                         children: [
-                          Expanded(child: const TelephoneWidget()),
-                          const SizedBox(width: 8),
-                          Expanded(child: const AddressWidget()),
+                          Expanded(child: TelephoneWidget()),
+                          SizedBox(width: 8),
+                          Expanded(child: AddressWidget()),
                         ],
                       ),
                       SizedBox(height: _verticalPadding),
-                      Consumer(builder: (context, watch, _) {
+                      Consumer(builder: (context, ref, _) {
                         return Row(
                           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                           children: [
-                            Expanded(child: const DepartmentDropdown()),
-                            if (watch(selectedDepartmentProvider).state != null)
-                              Expanded(
-                                  child: const LocalityDropdown(
-                                      isTwoColumn: true)),
+                            const Expanded(child: DepartmentDropdown()),
+                            if (ref.watch(selectedDepartmentProvider) != null)
+                              const Expanded(
+                                  child: LocalityDropdown(isTwoColumn: true)),
                           ],
                         );
                       }),
                       SizedBox(height: _verticalPadding),
                       const AuthorizationDate(),
                       SizedBox(height: _verticalPadding),
-                      Row(
+                      const Row(
                         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                         children: [
                           Expanded(
                               flex: 2,
-                              child: const SizedBox(
+                              child: SizedBox(
                                   width: 400,
-                                  child: const PreferredScheduleDropdown())),
-                          const SizedBox(width: 8),
+                                  child: PreferredScheduleDropdown())),
+                          SizedBox(width: 8),
                           Expanded(
                               flex: 1,
                               child: SizedBox(
-                                  width: 100, child: const SessionsWidget())),
+                                  width: 100, child: SessionsWidget())),
                         ],
                       ),
                       SizedBox(height: _verticalPadding),
