@@ -4,21 +4,24 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../models/models.dart';
 import '../../../providers/providers.dart';
 
-class IdNumberWidget extends StatefulWidget {
+class IdNumberWidget extends ConsumerStatefulWidget {
   const IdNumberWidget({Key? key, required this.isTwoColumn}) : super(key: key);
   final bool isTwoColumn;
 
   @override
-  _IdNumberWidgetState createState() => _IdNumberWidgetState();
+  ConsumerState<IdNumberWidget> createState() => _IdNumberWidgetState();
 }
 
-class _IdNumberWidgetState extends State<IdNumberWidget> {
+class _IdNumberWidgetState extends ConsumerState<IdNumberWidget> {
   late TextEditingController _controller;
 
   @override
   void initState() {
     _controller = TextEditingController.fromValue(
-        TextEditingValue(text: context.read(patientIdNumberProvider).state));
+      TextEditingValue(
+        text: ref.read(patientIdNumberProvider),
+      ),
+    );
     super.initState();
   }
 
@@ -30,32 +33,33 @@ class _IdNumberWidgetState extends State<IdNumberWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return Consumer(builder: (context, watch, _) {
-      TypeOfId? _typeOfId = watch(typeOfIdProvider).state;
-      if (_typeOfId != null) {
+    return Consumer(builder: (context, ref, _) {
+      TypeOfId? typeOfId = ref.watch(typeOfIdProvider);
+      if (typeOfId != null) {
         return !widget.isTwoColumn
             ? Column(
                 children: [
-                  SizedBox(height: 16),
+                  const SizedBox(height: 16),
                   TextFormField(
                     controller: _controller,
                     autovalidateMode: AutovalidateMode.onUserInteraction,
                     validator: (value) => value != null && value.length > 15
                         ? 'Máximo 15 caracteres'
                         : null,
-                    onChanged: (value) =>
-                        context.read(patientIdNumberProvider).state = value,
-                    keyboardType: _typeOfId.id != 5
+                    onChanged: (value) => ref
+                        .read(patientIdNumberProvider.notifier)
+                        .state = value,
+                    keyboardType: typeOfId.id != 5
                         ? TextInputType.number
                         : TextInputType.text,
                     textInputAction: TextInputAction.next,
-                    inputFormatters: _typeOfId.id != 5
+                    inputFormatters: typeOfId.id != 5
                         ? [
                             FilteringTextInputFormatter.allow(RegExp(r'[0-9]')),
                           ]
                         : [],
                     decoration: InputDecoration(
-                      contentPadding: EdgeInsets.symmetric(
+                      contentPadding: const EdgeInsets.symmetric(
                           horizontal: 20.0, vertical: 15.0),
                       labelText: 'Número de identidad*',
                       border: OutlineInputBorder(
@@ -66,7 +70,7 @@ class _IdNumberWidgetState extends State<IdNumberWidget> {
               )
             : Row(
                 children: [
-                  SizedBox(width: 8),
+                  const SizedBox(width: 8),
                   Expanded(
                     child: TextFormField(
                       controller: _controller,
@@ -74,20 +78,21 @@ class _IdNumberWidgetState extends State<IdNumberWidget> {
                       validator: (value) => value != null && value.length > 15
                           ? 'Máximo 15 caracteres'
                           : null,
-                      onChanged: (value) =>
-                          context.read(patientIdNumberProvider).state = value,
-                      keyboardType: _typeOfId.id != 5
+                      onChanged: (value) => ref
+                          .read(patientIdNumberProvider.notifier)
+                          .state = value,
+                      keyboardType: typeOfId.id != 5
                           ? TextInputType.number
                           : TextInputType.text,
                       textInputAction: TextInputAction.next,
-                      inputFormatters: _typeOfId.id != 5
+                      inputFormatters: typeOfId.id != 5
                           ? [
                               FilteringTextInputFormatter.allow(
                                   RegExp(r'[0-9]')),
                             ]
                           : [],
                       decoration: InputDecoration(
-                        contentPadding: EdgeInsets.symmetric(
+                        contentPadding: const EdgeInsets.symmetric(
                             horizontal: 20.0, vertical: 15.0),
                         labelText: 'Número de identidad*',
                         border: OutlineInputBorder(
@@ -98,7 +103,7 @@ class _IdNumberWidgetState extends State<IdNumberWidget> {
                 ],
               );
       } else {
-        return SizedBox();
+        return const SizedBox();
       }
     });
   }
