@@ -2,20 +2,20 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../providers/providers.dart';
 
-class NameWidget extends StatefulWidget {
+class NameWidget extends ConsumerStatefulWidget {
   const NameWidget({Key? key}) : super(key: key);
 
   @override
-  _NameWidgetState createState() => _NameWidgetState();
+  ConsumerState<NameWidget> createState() => _NameWidgetState();
 }
 
-class _NameWidgetState extends State<NameWidget> {
+class _NameWidgetState extends ConsumerState<NameWidget> {
   late TextEditingController _controller;
 
   @override
   void initState() {
     _controller = TextEditingController.fromValue(
-        TextEditingValue(text: context.read(patientNameProvider).state));
+        TextEditingValue(text: ref.read(patientNameProvider)));
     super.initState();
   }
 
@@ -30,13 +30,15 @@ class _NameWidgetState extends State<NameWidget> {
     return TextFormField(
       autofocus: true,
       controller: _controller,
-      onChanged: (value) => context.read(patientNameProvider).state = value,
+      onChanged: (value) =>
+          ref.read(patientNameProvider.notifier).state = value,
       autovalidateMode: AutovalidateMode.onUserInteraction,
       validator: (value) =>
           value != null && value.length > 50 ? 'Máximo 50 caracteres' : null,
       textInputAction: TextInputAction.next,
       decoration: InputDecoration(
-        contentPadding: EdgeInsets.symmetric(horizontal: 20.0, vertical: 15.0),
+        contentPadding:
+            const EdgeInsets.symmetric(horizontal: 20.0, vertical: 15.0),
         labelText: 'Nombre del paciente*',
         border: OutlineInputBorder(borderRadius: BorderRadius.circular(5.0)),
       ),
