@@ -9,15 +9,17 @@ class PreferredScheduleDropdown extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Consumer(
-      builder: (context, watch, _) {
-        List<String> _scheduleNames = schedules.map((e) => e.time).toList();
+      builder: (context, ref, _) {
+        final scheduleNames = schedules.map((e) => e.time).toList();
         return AppDropdownInput<String?>(
           hintText: 'Horario preferente*',
-          options: _scheduleNames,
-          value: watch(preferredScheduleProvider).state?.time,
+          options: scheduleNames,
+          value: ref.watch(preferredScheduleProvider)?.time,
           onChanged: (String? value) async {
-            context.read(preferredScheduleProvider).state =
-                schedules.firstWhere((element) => element.time == value);
+            ref.read(preferredScheduleProvider.notifier).state =
+                schedules.firstWhere(
+              (element) => element.time == value,
+            );
           },
           getLabel: (String? value) => value ?? 'Horario preferente*',
         );
