@@ -1,9 +1,24 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
 import '../../../providers/providers.dart';
-import '../../../models/models.dart';
-import '../../../services/services.dart';
 import '../../screens.dart';
+
+final _submitEnabledProvider = Provider.autoDispose((ref) {
+  return ref.watch(authotizationDateProvider) != null &&
+      ref.watch(typeOfTherapyProvider) != null &&
+      ref.watch(preferredScheduleProvider) != null &&
+      ref.watch(typeOfIdProvider) != null &&
+      ref.watch(numberOfTherapiesProvider) > 0 &&
+      ref.watch(patientNameProvider).trim().isNotEmpty &&
+      ref.watch(patientDiagnosisProvider).trim().isNotEmpty &&
+      ref.watch(patientIdNumberProvider).trim().isNotEmpty &&
+      ref.watch(telephoneNumberProvider) > 1000000 &&
+      ref.watch(patientAgeProvider) > 0 &&
+      ref.watch(addressNumberProvider).trim().isNotEmpty;
+});
 
 class SubmitButton extends StatelessWidget {
   const SubmitButton({
@@ -12,51 +27,44 @@ class SubmitButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Consumer(builder: (context, watch, _) {
-      var _canSend = watch(authotizationDateProvider).state != null &&
-          watch(typeOfTherapyProvider).state != null &&
-          watch(preferredScheduleProvider).state != null &&
-          watch(typeOfIdProvider).state != null &&
-          watch(numberOfTherapiesProvider).state > 0 &&
-          watch(patientNameProvider).state.trim().isNotEmpty &&
-          watch(patientDiagnosisProvider).state.trim().isNotEmpty &&
-          watch(patientIdNumberProvider).state.trim().isNotEmpty &&
-          watch(telephoneNumberProvider).state > 1000000 &&
-          watch(patientAgeProvider).state > 0 &&
-          watch(addressNumberProvider).state.trim().isNotEmpty;
+    return Consumer(builder: (context, ref, _) {
+      final canSend = ref.watch(_submitEnabledProvider);
       return ElevatedButton(
-        onPressed: _canSend
-            ? () async => await _validateAndSubmitResults(context)
+        onPressed: canSend
+            ? () async => await _validateAndSubmitResults(context, ref)
             : null,
-        child: Container(
-          padding: const EdgeInsets.symmetric(horizontal: 20.0, vertical: 15.0),
+        child: const Padding(
+          padding: EdgeInsets.symmetric(horizontal: 20.0, vertical: 15.0),
           child: Text('Solicitar cita'),
         ),
       );
     });
   }
 
-  Future<void> _validateAndSubmitResults(BuildContext context) async {
+  Future<void> _validateAndSubmitResults(
+    BuildContext context,
+    WidgetRef ref,
+  ) async {
     try {
-      context.read(isLoadingProvider).state = true;
-      Department? department = context.read(selectedDepartmentProvider).state;
-      Locality? locality = context.read(selectedLocalityProvider).state;
-      Schedule preferedSchedule =
-          context.read(preferredScheduleProvider).state!;
-      TypeOfId typeOfId = context.read(typeOfIdProvider).state!;
-      int numberOfTherapies = context.read(numberOfTherapiesProvider).state;
-      int telephone = context.read(telephoneNumberProvider).state;
-      String name = context.read(patientNameProvider).state.trim();
-      String diagnosis = context.read(patientDiagnosisProvider).state.trim();
-      String idNumber = context.read(patientIdNumberProvider).state.trim();
-      InsuranceCompany? insuranceCompany =
-          context.read(insuranceNameProvider).state;
-      int age = context.read(patientAgeProvider).state;
-      String address = context.read(addressNumberProvider).state.trim();
-      DateTime authDate = context.read(authotizationDateProvider).state!;
-      Therapy typeOfTherapy = context.read(typeOfTherapyProvider).state!;
+      ref.read(isLoadingProvider.notifier).state = true;
+      final mounted = context.mounted;
 
-      Map<String, dynamic> data = context.read(formServiceProvider).processData(
+      final department = ref.read(selectedDepartmentProvider);
+      final locality = ref.read(selectedLocalityProvider);
+      final preferedSchedule = ref.read(preferredScheduleProvider)!;
+      final typeOfId = ref.read(typeOfIdProvider)!;
+      final numberOfTherapies = ref.read(numberOfTherapiesProvider);
+      final telephone = ref.read(telephoneNumberProvider);
+      final name = ref.read(patientNameProvider).trim();
+      final diagnosis = ref.read(patientDiagnosisProvider).trim();
+      final idNumber = ref.read(patientIdNumberProvider).trim();
+      final insuranceCompany = ref.read(insuranceNameProvider);
+      final age = ref.read(patientAgeProvider);
+      final address = ref.read(addressNumberProvider).trim();
+      final authDate = ref.read(authotizationDateProvider)!;
+      final typeOfTherapy = ref.read(typeOfTherapyProvider)!;
+
+      final data = ref.read(formServiceProvider).processData(
             locality: locality,
             department: department,
             name: name,
@@ -72,80 +80,91 @@ class SubmitButton extends StatelessWidget {
             authDate: authDate,
             preferedSchedule: preferedSchedule,
           );
-      int statusEmail = await context
-          .read(formServiceProvider)
-          .submitForm(data: data, type: SubmitSheetsOrEmailFunction.email);
-      int statusSheets = await context
-          .read(formServiceProvider)
-          .submitForm(data: data, type: SubmitSheetsOrEmailFunction.sheets);
+      final responses = await ref.read(formServiceProvider).submitForm(
+            data: data,
+          );
+      final statusEmail = responses.emailResultStatusCode;
+      final statusSheets = responses.sheetsResultStatusCode;
       if (statusEmail == 200 && statusSheets == 200) {
-        context.read(isLoadingProvider).state = false;
-        showDialog(
-          context: context,
-          barrierDismissible: false,
-          builder: (context) {
-            return FormSubmittedDialog();
-          },
+        if (!mounted) return;
+        unawaited(
+          showDialog(
+            context: context,
+            barrierDismissible: false,
+            builder: (context) {
+              return const FormSubmittedDialog();
+            },
+          ),
         );
       }
     } catch (e) {
       debugPrint(e.toString());
-      context.read(isLoadingProvider).state = false;
-      showDialog(
-        context: context,
-        barrierDismissible: false,
-        builder: (context) {
-          return AlertDialog(
-            title: Text('Error'),
-            content: Text('$e'),
-            actions: [
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.pop(context);
-                },
-                child: Text('Cerrar'),
-              ),
-            ],
-          );
-        },
+      ref.read(isLoadingProvider.notifier).state = false;
+      unawaited(
+        showDialog(
+          context: context,
+          barrierDismissible: false,
+          builder: (context) {
+            return AlertDialog(
+              title: const Text('Error'),
+              content: Text('$e'),
+              actions: [
+                ElevatedButton(
+                  onPressed: () {
+                    Navigator.pop(context);
+                  },
+                  child: const Text('Cerrar'),
+                ),
+              ],
+            );
+          },
+        ),
       );
+    } finally {
+      ref.read(isLoadingProvider.notifier).state = false;
     }
   }
 }
 
-class FormSubmittedDialog extends StatelessWidget {
+class FormSubmittedDialog extends ConsumerWidget {
   const FormSubmittedDialog({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     return AlertDialog(
-      title: Text('Solicitud realizada'),
-      content: Text(
-          'Muy pronto nos pondremos en contacto contigo para confirmar tu agendamiento'),
+      title: const Text('Solicitud realizada'),
+      content: const Text(
+        'Muy pronto nos pondremos en contacto contigo para confirmar tu agendamiento',
+      ),
       actions: [
         TextButton(
-          onPressed: () => _resetForm(context),
-          child: Text('Okay'),
+          onPressed: () => _resetForm(context, ref),
+          child: const Text('Okay'),
         ),
       ],
     );
   }
 
-  void _resetForm(BuildContext context) {
-    context.read(selectedDepartmentProvider).state = null;
-    context.read(selectedLocalityProvider).state = null;
-    context.read(preferredScheduleProvider).state = null;
-    context.read(typeOfIdProvider).state = null;
-    context.read(insuranceNameProvider).state = null;
-    context.read(authotizationDateProvider).state = null;
-    context.read(typeOfTherapyProvider).state = null;
-    context.read(numberOfTherapiesProvider).state = 0;
-    context.read(telephoneNumberProvider).state = 0;
-    context.read(patientNameProvider).state = '';
-    context.read(patientDiagnosisProvider).state = '';
-    context.read(patientIdNumberProvider).state = '';
-    context.read(patientAgeProvider).state = 0;
-    context.read(addressNumberProvider).state = '';
-    Navigator.pushReplacementNamed(context, FormScreen.route);
+  void _resetForm(
+    BuildContext context,
+    WidgetRef ref,
+  ) {
+    ref.read(selectedDepartmentProvider.notifier).state = null;
+    ref.read(selectedLocalityProvider.notifier).state = null;
+    ref.read(preferredScheduleProvider.notifier).state = null;
+    ref.read(typeOfIdProvider.notifier).state = null;
+    ref.read(insuranceNameProvider.notifier).state = null;
+    ref.read(authotizationDateProvider.notifier).state = null;
+    ref.read(typeOfTherapyProvider.notifier).state = null;
+    ref.read(numberOfTherapiesProvider.notifier).state = 0;
+    ref.read(telephoneNumberProvider.notifier).state = 0;
+    ref.read(patientNameProvider.notifier).state = '';
+    ref.read(patientDiagnosisProvider.notifier).state = '';
+    ref.read(patientIdNumberProvider.notifier).state = '';
+    ref.read(patientAgeProvider.notifier).state = 0;
+    ref.read(addressNumberProvider.notifier).state = '';
+    unawaited(
+      Navigator.pushReplacementNamed(context, FormScreen.route),
+    );
   }
 }
